+++ custom_components/sector/lock.py
@@ -1,5 +1,5 @@
 """Adds Lock for Sector integration."""
-import logging
+from __future__ import annotations
 
 from homeassistant.components.lock import LockEntity, LockEntityDescription
 from homeassistant.config_entries import ConfigEntry
@@ -12,10 +12,8 @@ from homeassistant.helpers.update_coordinator import (
     DataUpdateCoordinator,
 )
 
-from .__init__ import SectorAlarmHub
 from .const import CONF_CODE, CONF_CODE_FORMAT, CONF_LOCK, DOMAIN
-
-_LOGGER = logging.getLogger(__name__)
+from .coordinator import SectorAlarmHub
 
 
 async def async_setup_entry(
