+++ custom_components/sector/sensor.py
@@ -1,5 +1,5 @@
 """Adds Temp sensors for Sector integration."""
-import logging
+from __future__ import annotations
 
 from homeassistant.components.sensor import (
     SensorDeviceClass,
@@ -17,10 +17,8 @@ from homeassistant.helpers.update_coordinator import (
     DataUpdateCoordinator,
 )
 
-from .__init__ import SectorAlarmHub
 from .const import CONF_TEMP, DOMAIN
-
-_LOGGER = logging.getLogger(__name__)
+from .coordinator import SectorAlarmHub
 
 
 async def async_setup_entry(
